bot/requirements.txt
anthropic>=0.97.0
httpx>=0.28.1
fastapi>=0.136.0
uvicorn[standard]>=0.46.0
python-dotenv>=1.2.0
pyyaml>=6.0
